+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+
 * [Enhance] Enhance err logging and response examples
+
+## 0.7.0 / 2019-06-28
 * [FEATURE] Add 'target' label, its value is ipaddress
 
 ## 0.6.0 / 2019-06-21
+
 * [CHANGE] Change type of token log form INFO to DEBUG.
+
 ## 0.5.0 / 2019-06-15
 
 ### **Breaking changes**